+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - changed: set autolock on newly created GP object
 - changed: add first keyframe on current frame when creating object (as native behavior)
 - fixed: add undo step at draw object creation
+- fixed: refresh all viewport areas when toggling toolbar
 
 1.8.7